extract op matching helpers
